add histogram of NN
In main.py the train result include training_history, sample of net_y and sample of true_y (at same x). The two samples are saved as vectors in text files. In solver.py, self.hist_size denotes the number of sample points. After the last step of training, the two samples are computed and are saved as part of return of model.train(). Build_true() still work because I set new sample 0.